move rollupMomentPlugin.ts out from src
