+++ .eslintrc.js
@@ -78,7 +78,8 @@ module.exports = {
         "**/test/**",
         "**/vite.config.ts",
         "**/vitest.config.ts",
-        "**/.eslintrc.js"
+        "**/.eslintrc.js",
+        "**/rollupMomentPlugin.ts"
       ],
     }],
     // Sort imports: React first, then npm packages, then local files, then CSS.
